+++ CHANGELOG.md
@@ -11,9 +11,11 @@ For a full diff see [`f4208b3...main`][f4208b3...main].
 ### Added
 
 - Added `ValueGenerator` ([#1]), by [@localheinz]
+- Added `OptionalValueGenerator` ([#2]), by [@localheinz]
 
 [f4208b3...main]: https://github.com/ergebnis/data-generator/compare/f4208b3...main
 
 [#1]: https://github.com/ergebnis/data-generator/pull/1
+[#2]: https://github.com/ergebnis/data-generator/pull/2
 
 [@localheinz]: https://github.com/localheinz